replaced float32 with float16 where feasible

+++ src/data_loader.py
@@ -16,7 +16,7 @@ def classification_data_loader(df_x, df_y, dataset_path, batch_size=32):
                 image_path = os.path.join(dataset_path, f"{row['ImageId']}")
                 image = tf.io.read_file(image_path)
                 image = tf.image.decode_jpeg(image, channels=3)
-                image = tf.cast(image, tf.float32) / 255.0
+                image = tf.cast(image, tf.float16) / 255.0
                 images.append(image)
                 labels.append(row['HasShip'])
 
@@ -40,7 +40,7 @@ def classification_validation_data_loader(df_x, df_y, dataset_path, batch_size=3
                 image_path = os.path.join(dataset_path, f"{row['ImageId']}")
                 image = tf.io.read_file(image_path)
                 image = tf.image.decode_jpeg(image, channels=3)
-                image = tf.cast(image, tf.float32) / 255.0
+                image = tf.cast(image, tf.float16) / 255.0
                 images.append(image)
                 labels.append(row['HasShip'])
             yield tf.stack(images), tf.stack(labels)
